hack: Append rather than prepend buildEnv path environment variables
When one has a package of name X in a Nix development environment (or
nix-shell) and a package of the same name appears in a local
workspace, previous implementation ignored the local package. With
this change, the local package overrides the package from Nix. If
local workspace has no package named X, the one from Nix will be
visible.

Theoretically, it should be possible to achieve the same without
modifying nix-ros-overlay, by just disabling wrapping of
binaries (e.g. ros2) in the build-env with `wrapPrograms = false;`.
But for some reason, it doesn't work. It makes some packages not
visible at all, which looks like a bug. Until this bug is fixed, we
just do what's in this commit.

It's not sufficient to change only AMENT_PREFIX_PATH. It seems that
`ros launch` searches the packages not based on
AMENT_PREFIX_PATH (modified by the last commit), but based on
PYTHONPATH. Therefore, changes performed in Python code in a local
workspace launched via `ros launch` were not visible, because `ros
launch` loads the node from Nix package and not the one from local
workspace.

We fix that by suffixing also PYTHONPATH. And while we're at it, we
also change the remaining variables. It seems to be the right thing
for our use case.