hack: Append rather than prepend build-env path to AMENT_PREFIX_PATH
When one has a package of name X in a Nix development environment (or
nix-shell) and a package of the same name appears in a local
workspace, previous implementation ignored the local package. With
this change, the local package overrides the package from Nix. If
local workspace has no package named X, the one from Nix will be
visible.

Theoretically, it should be possible to achieve the same without
modifying nix-ros-overlay, by just disabling wrapping of
binaries (e.g. ros2) in the build-env with `wrapPrograms = false;`.
But for some reason, it doesn't work. It makes some packages not
visible at all, which looks like a bug. Until this bug is fixed, we
just do what's in this commit.