hack: Append rather than prepend build-env path to other environment variables
It turns out, that the previous commit was not sufficient. It seems
that `ros launch` searches the packages not based on
AMENT_PREFIX_PATH (modified by the last commit), but based on
PYTHONPATH. Therefore, changes performed in Python code in a local
workspace launched via `ros launch` were not visible, because `ros
launch` loads the node from Nix package and not the one from local
workspace.

We fix that by suffixing also PYTHONPATH. And while we're at it, we
also change the remaining variables. It seems to be the right thing
for our use case.