Add GitHub Actions workflow to record history of the certicar branch
The certicar branch is periodically rebased. To keep history across
rebases, this commit ensures that after every push, the state of the
branch is recorded in the certicar-history branch. Since that branch
will never be rebased, the history will remain there.

Before this commit, we did the same, but manually.