Add flake template
With this change, one can initialize a ROS project by running:

    nix flake init --template github:lopsided98/nix-ros-overlay

Besides adding the template, this commit also updates the README to
list the same template as the one from the flake. The listing in the
README can be easily updated to match the flake.nix file by running
the "mdsh" tool.