flake.nix: Expose all top-level packages
Previously, the legacyPackages output contained only the content of
the rosPackages attribute, but things like colcon and gazebo were not
available there.

With this change, we expose all packages defined in pkgs/default.nix
as well as the rosPackages attribute. This is the same as when using
just default.nix without flakes. To maintain backward compatibility,
we still keep ROS distro package sets at the top level. This means,
that the same ROS package is available at two places. For example,
ros-core from humble is available as:

- .#humble.ros-core (before and after this commit)
- .#rosPackages.humble.ros-core (after this commit)