mrpt: Patch vendored source code of MRPT packages
MRPT ROS packaging was recently split to multiple smaller ROS
packages, but the source code of all of them is still maintained in
the original big repo. The new packages fetch the original repo and
build just a part of it. Obviously, with Nix, we have to fetch source
code ourselves, instead of letting CMake do this.