Partial support of finalAttrs pattern for buildRosPackage
With this change, it's possible to use finalAttrs pattern with
buildRosPackage function. Unfortunately, it works only for non-Python
packages.

This allows to define ROS packages as follows:

    buildRosPackage (finalAttrs: {
      pname = "ros-${finalAttrs.passthru.rosDistro}-package";
      version = "1.0.0";

      src = ...;

      # ...
    })

The advantage is that if a single Nix expression is built for
different ROS distros, the name of the derivation shows for which ROS
distribution it was built.

The same change would also allow expressing conditional dependencies
in package.xml in Nix expressions. With it, there would be one-to-one
mapping between package.xml and corresponding Nix expression.

This commit tries to make minimal changes to existing code. However,
in the future, it would be beneficial to refactor it to be simpler and
easier to read. For example, Python handling in updateAttrsForRos
could be moved to a separate function.