Update patchAmentVendorGit to use automatically extracted information about vendored source
The information is store in vendored-source.json files in the package
directories. The file can be automatically generated by running, e.g.:

    $(nix-build -A rosPackages.jazzy.gz-math-vendor.updateAmentVendor)'

This command executes cmake for the given packages with an alternative
version of the ament_vendor macro, which generates the file. The above
command generates:

    $ cat distros/jazzy/gz-math-vendor/vendored-source.json
    {
      "url": "https://github.com/gazebosim/gz-math.git",
      "rev": "gz-math7_7.5.2",
      "hash": "sha256-LwYeyv8nwX06n5ju+ra2uqNMedMSLRumem8qDHXtNns="
    }

patchAmentVendorGit is updated to pick the information about patching
from that file instead of from its arguments. patchGzAmentVendorGit is
updated similarly.