cargo-ament-build: Build also as a Python package
Python package is needed as a dependency for colcon-ros-cargo. The
Python package doesn't contain any Python module, it's just a
cargo-ament-build binary wrapped into a Python wheel.

In order to generate the Python wheel, we change how we build the
package. Instead of using expressions generated by crate2nix (which
cannot generate the wheel, or at least I don't know how), we use the
standard nixpkgs way, which utilizes cargoSetupHook and
maturinSetupHook.